Hooked up Mimbuild to FOFC.
