+++ tetrad-lib/src/main/java/edu/cmu/tetrad/search/FindOneFactorClusters.java
@@ -21,153 +21,24 @@
 
 package edu.cmu.tetrad.search;
 
-import edu.cmu.tetrad.data.CovarianceMatrix;
-import edu.cmu.tetrad.data.DataModel;
-import edu.cmu.tetrad.data.DataSet;
-import edu.cmu.tetrad.data.ICovarianceMatrix;
+import edu.cmu.tetrad.data.*;
 import edu.cmu.tetrad.graph.*;
 import edu.cmu.tetrad.sem.*;
-import edu.cmu.tetrad.util.ChoiceGenerator;
-import edu.cmu.tetrad.util.TetradLogger;
+import edu.cmu.tetrad.util.*;
 
 import java.util.*;
+
+import static java.lang.Math.abs;
+import static java.lang.Math.sqrt;
 
 
 /**
  * Implements FindOneFactorCluster by Erich Kummerfeld (adaptation of a two factor
- * sextet algorithm to a one factor tetrad algorithm).
+ * quartet algorithm to a one factor tetrad algorithm).
  *
  * @author Joseph Ramsey
  */
 public class FindOneFactorClusters {
-
-    private boolean extraShuffle;
-
-    public enum SortKey {size, pValue}
-
-    public enum Algorithm {SAG, GAP}
-    public enum AlgType {strict, lax, laxWithSpeedup}
-
-    private ICovarianceMatrix cov;
-
-    // The list of all variables.
-    private List<Node> variables;
-
-    private Algorithm algorithm = Algorithm.GAP;
-//    private Algorithm algorithm = Algorithm.SAG;
-
-    // The type of the algorithm
-    private AlgType algType = AlgType.laxWithSpeedup;
-
-    // The tetrad test--using Ricardo's. Used only for Wishart.
-    private TetradTest test;
-
-    // The significance level for the tetrad test.
-    private double alpha;
-
-    // The minimum p vaue for output clusters; clusters with lower p values will be ignored.
-    private Double clusterMinP = Double.NaN;
-
-    // Wishart or Bollen.
-    private TestType testType;
-
-    // The Bollen test. Testing two tetrads simultaneously.
-    private DeltaTetradTest deltaTest;
-
-    // independence test.
-    private IndependenceTest indTest;
-
-    private transient DataModel dataModel;
-
-    // The depth of the PC search, -2 if the PC search should not be run.
-    private List<List<Node>> clusters = new ArrayList<List<Node>>();
-
-    private int depth = 0; // -2 to turn PC off
-
-    private SortKey sortKey = SortKey.pValue;
-
-    private boolean verbose = false;
-
-    //========================================PUBLIC METHODS====================================//
-
-    public FindOneFactorClusters(ICovarianceMatrix cov, TestType testType, double alpha) {
-        this.variables = cov.getVariables();
-        this.test = new ContinuousTetradTest(cov, testType, alpha);
-        this.indTest = new IndTestFisherZ(cov, alpha);
-        this.alpha = alpha;
-        this.testType = testType;
-        deltaTest = new DeltaTetradTest(cov);
-        this.dataModel = cov;
-        this.cov = cov;
-    }
-
-    public FindOneFactorClusters(DataSet dataSet, TestType testType, double alpha) {
-        if (dataSet.isContinuous()) {
-            this.variables = dataSet.getVariables();
-            this.test = new ContinuousTetradTest(dataSet, testType, alpha);
-            this.indTest = new IndTestFisherZ(dataSet, alpha);
-            this.alpha = alpha;
-            this.testType = testType;
-            this.dataModel = dataSet;
-
-            if (testType == TestType.TETRAD_DELTA) {
-                deltaTest = new DeltaTetradTest(dataSet);
-//                deltaTest.setCacheFourthMoments(false);
-            }
-
-            this.cov = new CovarianceMatrix(dataSet);
-        } else if (dataSet.isDiscrete()) {
-            this.variables = dataSet.getVariables();
-            this.test = new DiscreteTetradTest(dataSet, alpha);
-            this.indTest = new IndTestChiSquare(dataSet, alpha);
-            this.alpha = alpha;
-            this.testType = testType;
-            this.dataModel = dataSet;
-
-            if (testType == TestType.TETRAD_DELTA) {
-                deltaTest = new DeltaTetradTest(dataSet);
-//                deltaTest.setCacheFourthMoments(false);
-            }
-        }
-    }
-
-    public Graph search() {
-        TetradLogger.getInstance().log("info", "FOFC alpha = " + alpha + " test = " + testType);
-
-        long start = System.currentTimeMillis();
-
-        Set<Set<Integer>> clusters;
-
-        if (algorithm == FindOneFactorClusters.Algorithm.GAP) {
-            clusters = estimateClustersGAP(); // Triples first.
-        } else if (algorithm == FindOneFactorClusters.Algorithm.SAG) {
-            clusters = estimateClustersSAG(); // Quartets first.
-        } else {
-            throw new IllegalStateException();
-        }
-
-        for (Set<Integer> cluster : clusters) {
-            this.clusters.add(variablesForIndices(new ArrayList<Integer>(cluster)));
-        }
-
-        ClusterUtils.logClusters(clusters, variables);
-
-        long stop = System.currentTimeMillis();
-
-        long elapsed = stop - start;
-
-        TetradLogger.getInstance().log("elapsed", "Elapsed " + elapsed + " ms");
-
-        return convertToGraph(clusters);
-    }
-
-    public boolean isVerbose() {
-        return verbose;
-    }
-
-    public void setVerbose(boolean verbose) {
-        this.verbose = verbose;
-    }
 
     public Algorithm getAlgorithm() {
         return algorithm;
@@ -177,84 +48,255 @@ public class FindOneFactorClusters {
         this.algorithm = algorithm;
     }
 
-    public void setDepth(int depth) {
-        this.depth = depth;
+    public enum Algorithm {SAG, GAP}
+
+    private CorrelationMatrix corr;
+    // The list of all variables.
+    private List<Node> variables;
+
+    // The significance level.
+    private double alpha;
+
+    private TestType testType = TestType.TETRAD_DELTA;
+
+    // The Delta test. Testing two tetrads simultaneously.
+    private DeltaTetradTest test;
+
+    // The tetrad test--using Ricardo's. Used only for Wishart.
+    private ContinuousTetradTest test2;
+
+    // The data.
+    private transient DataModel dataModel;
+
+    private List<List<Node>> clusters;
+
+    private int depth = 0;
+    private boolean verbose = false;
+    private boolean significanceCalculated = false;
+    private Algorithm algorithm = Algorithm.GAP;
+
+    //========================================PUBLIC METHODS====================================//
+
+    public FindOneFactorClusters(ICovarianceMatrix cov, TestType testType, Algorithm algorithm, double alpha) {
+        if (testType == null) throw new NullPointerException("Null test type.");
+        cov = new CovarianceMatrix(cov);
+        this.variables = cov.getVariables();
+        this.alpha = alpha;
+        this.testType = testType;
+        this.test = new DeltaTetradTest(cov);
+        this.test2 = new ContinuousTetradTest(cov, testType, alpha);
+        this.dataModel = cov;
+        this.algorithm = algorithm;
+
+        this.corr = new CorrelationMatrix(cov);
+
+
     }
 
-    public void setExtraShuffle(boolean extraShuffle) {
-        this.extraShuffle = extraShuffle;
+    public FindOneFactorClusters(DataSet dataSet, TestType testType, Algorithm algorithm, double alpha) {
+        if (testType == null) throw new NullPointerException("Null test type.");
+        this.variables = dataSet.getVariables();
+        this.alpha = alpha;
+        this.testType = testType;
+        this.test = new DeltaTetradTest(dataSet);
+        this.test2 = new ContinuousTetradTest(dataSet, testType, alpha);
+        this.dataModel = dataSet;
+        this.algorithm = algorithm;
+
+        this.corr = new CorrelationMatrix(dataSet);
     }
 
-    public boolean isExtraShuffle() {
-        return extraShuffle;
+    // renjiey
+    private int findFrequentestIndex(Integer outliers[]) {
+        Map<Integer, Integer> map = new HashMap<Integer, Integer>();
+
+        for (int i = 0; i < outliers.length; i++) {
+            if (map.containsKey(outliers[i])) {
+                map.put(outliers[i], map.get(outliers[i]) + 1);
+            } else {
+                map.put(outliers[i], 1);
+            }
+        }
+
+        Set<Map.Entry<Integer, Integer>> set = map.entrySet();
+        Iterator<Map.Entry<Integer, Integer>> it = set.iterator();
+        int nums = 0;// how many times variable occur
+        int key = 0;// the number occur the most times
+
+        while (it.hasNext()) {
+            Map.Entry<Integer, Integer> entry = it.next();
+            if (entry.getValue() > nums) {
+                nums = entry.getValue();
+                key = entry.getKey();
+            }
+        }
+
+        return (key);
     }
 
+    // This is the main function. It remove variables in the data such that the remaining correlation matrix
+    // does not contain extreme value
+    // Inputs: correlation matrix, upper and lower bound for unacceptable correlations
+    // Output: and dynamic array of removed variables
+    // renjiey
+    private ArrayList<Integer> removeVariables(TetradMatrix correlationMatrix, double lowerBound, double upperBound,
+                                               double percentBound) {
+        Integer outlier[] = new Integer[correlationMatrix.rows() * (correlationMatrix.rows() - 1)];
+        int count = 0;
+        for (int i = 2; i < (correlationMatrix.rows() + 1); i++) {
+            for (int j = 1; j < i; j++) {
+
+                if ((Math.abs(correlationMatrix.get(i - 1, j - 1)) < lowerBound)
+                        || (Math.abs(correlationMatrix.get(i - 1, j - 1)) > upperBound)) {
+                    outlier[count * 2] = i;
+                    outlier[count * 2 + 1] = j;
+
+                } else {
+                    outlier[count * 2] = 0;
+                    outlier[count * 2 + 1] = 0;
+                }
+                count = count + 1;
+            }
+        }
+
+        //find out the variables that should be deleted
+        ArrayList<Integer> removedVariables = new ArrayList<Integer>();
+
+        // Added the percent bound jdramsey
+        while (outlier.length > 1 && removedVariables.size() < percentBound * correlationMatrix.rows()) {
+            //find out the variable that occurs most frequently in outlier
+            int worstVariable = findFrequentestIndex(outlier);
+            if (worstVariable > 0) {
+                removedVariables.add(worstVariable);
+            }
+
+            //remove the correlations having the bad variable (change the relevant variables to 0)
+            for (int i = 1; i < outlier.length + 1; i++) {
+                if (outlier[i - 1] == worstVariable) {
+                    outlier[i - 1] = 0;
+
+                    if (i % 2 != 0) {
+                        outlier[i] = 0;
+                    } else {
+                        outlier[i - 2] = 0;
+                    }
+                }
+            }
+
+            //delete zero elements in outlier
+            outlier = removeZeroIndex(outlier);
+        }
+
+        log(removedVariables.size() + " variables removed: " + variablesForIndices(removedVariables), true);
+
+        return (removedVariables);
+    }
+
+    // renjiey
+    private Integer[] removeZeroIndex(Integer outlier[]) {
+        List<Integer> list = new ArrayList<Integer>();
+        for (int i = 0; i < outlier.length; i++) {
+            list.add(outlier[i]);
+        }
+        for (Integer element : outlier) {
+            if (element < 1) {
+                list.remove(element);
+            }
+        }
+        return list.toArray(new Integer[1]);
+    }
+
+
+    public Graph search() {
+        Set<List<Integer>> allClusters;
+
+        if (algorithm == Algorithm.SAG) {
+            allClusters = estimateClustersTetradsFirst();
+        } else if (algorithm == Algorithm.GAP) {
+            allClusters = estimateClustersTriplesFirst();
+        } else {
+            throw new IllegalStateException("Expected SAG or GAP: " + testType);
+        }
+        this.clusters = variablesForIndices2(allClusters);
+        return convertToGraph(allClusters);
+    }
 
     //========================================PRIVATE METHODS====================================//
 
-    private Set<Set<Integer>> estimateClustersGAP() {
-        Set<Set<Integer>> threeClusters = findThreeClusters();
-        return combineThreeClusters(threeClusters);
+    // This is the main algorithm.
+    private Set<List<Integer>> estimateClustersTriplesFirst() {
+//        List<Integer> _variables = new ArrayList<Integer>();
+//        for (int i = 0; i < variables.size(); i++) _variables.add(i);
+        List<Integer> _variables = allVariables();
+
+        Set<Set<Integer>> triples = findPuretriples(_variables);
+        Set<Set<Integer>> combined = combinePuretriples(triples, _variables);
+
+        Set<List<Integer>> _combined = new HashSet<List<Integer>>();
+
+        for (Set<Integer> c : combined) {
+            List a = new ArrayList<Integer>(c);
+            Collections.sort(a);
+            _combined.add(a);
+        }
+
+        return _combined;
 
     }
 
-    private Set<Set<Integer>> findThreeClusters() {
-        if (variables.size() < 4) {
+    private List<Integer> allVariables() {
+        List<Integer> _variables = new ArrayList<Integer>();
+        for (int i = 0; i < variables.size(); i++) _variables.add(i);
+        return _variables;
+    }
+
+    private Set<List<Integer>> estimateClustersTetradsFirst() {
+        System.out.println("A");
+
+        List<Integer> _variables = allVariables();
+
+        Set<List<Integer>> pureClusters = findPureClusters(_variables);
+        Set<List<Integer>> mixedClusters = findMixedClusters(pureClusters, _variables, unionPure(pureClusters));
+        Set<List<Integer>> allClusters = new HashSet<List<Integer>>(pureClusters);
+        allClusters.addAll(mixedClusters);
+        return allClusters;
+
+    }
+
+    private Set<Set<Integer>> findPuretriples(List<Integer> allVariables) {
+        if (allVariables.size() < 4) {
             return new HashSet<Set<Integer>>();
         }
-
-        Map<Node, Set<Node>> adjacencies;
-
-        if (depth == -2) {
-            adjacencies = new HashMap<Node, Set<Node>>();
-
-            for (Node node : variables) {
-                HashSet<Node> _nodes = new HashSet<Node>(variables);
-                _nodes.remove(node);
-                adjacencies.put(node, _nodes);
-            }
-        } else {
-//            System.out.println("Running PC adjacency search...");
-            Graph graph = new EdgeListGraph(variables);
-            Fas fas = new Fas(graph, indTest);
-            fas.setVerbose(false);
-            fas.setDepth(depth);     // 1?
-            adjacencies = fas.searchMapOnly();
-//            System.out.println("...done.");
-        }
-
-        List<Integer> allVariables = new ArrayList<Integer>();
-        for (int i = 0; i < variables.size(); i++) allVariables.add(i);
 
         log("Finding pure triples.", true);
 
         ChoiceGenerator gen = new ChoiceGenerator(allVariables.size(), 3);
         int[] choice;
-        Set<Set<Integer>> fiveClusters = new HashSet<Set<Integer>>();
+        Set<Set<Integer>> puretriples = new HashSet<Set<Integer>>();
         CHOICE:
         while ((choice = gen.next()) != null) {
             int n1 = allVariables.get(choice[0]);
             int n2 = allVariables.get(choice[1]);
             int n3 = allVariables.get(choice[2]);
 
-            Set<Integer> triple = triple(n1, n2, n3);
+            List<Integer> triple = triple(n1, n2, n3);
 
-            if (!clique(triple, adjacencies)) {
-                continue;
-            }
+            if (zeroCorr(triple)) continue;
 
             for (int o : allVariables) {
                 if (triple.contains(o)) {
                     continue;
                 }
 
-                Set<Integer> cluster = quartet(n1, n2, n3, o);
+                List<Integer> quartet = quartet(n1, n2, n3, o);
 
-                if (!(quartetVanishes(cluster))) {
+                boolean vanishes = vanishes(quartet);
+
+                if (!vanishes) {
                     continue CHOICE;
                 }
 
-//                if (!(avgSumLnP(cluster) > -20)) {
+//                if (!(avgSumLnP(quartet) > -20)) {
 //                    continue CHOICE;
 //                }
             }
@@ -262,40 +304,32 @@ public class FindOneFactorClusters {
             HashSet<Integer> _cluster = new HashSet<Integer>(triple);
 
             if (verbose) {
-                log("++" + variablesForIndices(new ArrayList<Integer>(triple)), false);
+                log("++" + variablesForIndices(triple), false);
             }
 
-            fiveClusters.add(_cluster);
+            puretriples.add(_cluster);
         }
 
-        return fiveClusters;
+        return puretriples;
     }
 
-    private Set<Set<Integer>> combineThreeClusters(Set<Set<Integer>> threeClusters) {
+    private Set<Set<Integer>> combinePuretriples(Set<Set<Integer>> puretriples, List<Integer> _variables) {
         log("Growing pure triples.", true);
         Set<Set<Integer>> grown = new HashSet<Set<Integer>>();
 
-        List<Integer> _variables = new ArrayList<Integer>();
-        for (int i = 0; i < variables.size(); i++) _variables.add(i);
-
-
         // Lax grow phase with speedup.
-        if (algType == AlgType.lax) {
+        if (true) {
             Set<Integer> t = new HashSet<Integer>();
             int count = 0;
-            int total = threeClusters.size();
+            int total = puretriples.size();
 
             do {
-                if (!threeClusters.iterator().hasNext()) {
+                if (!puretriples.iterator().hasNext()) {
                     break;
                 }
 
-                Set<Integer> cluster = threeClusters.iterator().next();
+                Set<Integer> cluster = puretriples.iterator().next();
                 Set<Integer> _cluster = new HashSet<Integer>(cluster);
-
-                if (extraShuffle) {
-                    Collections.shuffle(_variables);
-                }
 
                 for (int o : _variables) {
                     if (_cluster.contains(o)) continue;
@@ -308,15 +342,12 @@ public class FindOneFactorClusters {
                     int[] choice;
 
                     while ((choice = gen.next()) != null) {
-                        int n1 = _cluster2.get(choice[0]);
-                        int n2 = _cluster2.get(choice[1]);
-
                         t.clear();
-                        t.add(n1);
-                        t.add(n2);
+                        t.add(_cluster2.get(choice[0]));
+                        t.add(_cluster2.get(choice[1]));
                         t.add(o);
 
-                        if (!threeClusters.contains(t)) {
+                        if (!puretriples.contains(t)) {
                             rejected++;
                         } else {
                             accepted++;
@@ -349,28 +380,24 @@ public class FindOneFactorClusters {
                     t.add(n2);
                     t.add(n3);
 
-                    threeClusters.remove(t);
+                    puretriples.remove(t);
                 }
 
                 if (verbose) {
                     System.out.println("Grown " + (++count) + " of " + total + ": " + variablesForIndices(new ArrayList<Integer>(_cluster)));
                 }
                 grown.add(_cluster);
-            } while (!threeClusters.isEmpty());
+            } while (!puretriples.isEmpty());
         }
 
         // Lax grow phase without speedup.
-        if (algType == AlgType.laxWithSpeedup) {
+        if (false) {
             int count = 0;
-            int total = threeClusters.size();
+            int total = puretriples.size();
 
             // Optimized lax version of grow phase.
-            for (Set<Integer> cluster : new HashSet<Set<Integer>>(threeClusters)) {
+            for (Set<Integer> cluster : new HashSet<Set<Integer>>(puretriples)) {
                 Set<Integer> _cluster = new HashSet<Integer>(cluster);
-
-                if (extraShuffle) {
-                    Collections.shuffle(_variables);
-                }
 
                 for (int o : _variables) {
                     if (_cluster.contains(o)) continue;
@@ -378,34 +405,39 @@ public class FindOneFactorClusters {
                     List<Integer> _cluster2 = new ArrayList<Integer>(_cluster);
                     int rejected = 0;
                     int accepted = 0;
-//
-                    ChoiceGenerator gen = new ChoiceGenerator(_cluster2.size(), 2);
+
+                    ChoiceGenerator gen = new ChoiceGenerator(_cluster2.size(), 4);
                     int[] choice;
 
                     while ((choice = gen.next()) != null) {
                         int n1 = _cluster2.get(choice[0]);
                         int n2 = _cluster2.get(choice[1]);
+                        int n3 = _cluster2.get(choice[2]);
+                        int n4 = _cluster2.get(choice[3]);
 
-                        Set<Integer> triple = triple(n1, n2, o);
+                        List<Integer> triple = triple(n1, n2, o);
 
-                        if (!threeClusters.contains(triple)) {
+                        Set<Integer> t = new HashSet<Integer>(triple);
+
+                        if (!puretriples.contains(t)) {
                             rejected++;
                         } else {
                             accepted++;
                         }
+
+//                        if (avgSumLnP(triple) < -10) continue CLUSTER;
                     }
-//
+
                     if (rejected > accepted) {
                         continue;
                     }
 
-//                    System.out.println("Adding " + o  + " to " + cluster);
                     _cluster.add(o);
                 }
 
-                for (Set<Integer> c : new HashSet<Set<Integer>>(threeClusters)) {
+                for (Set<Integer> c : new HashSet<Set<Integer>>(puretriples)) {
                     if (_cluster.containsAll(c)) {
-                        threeClusters.remove(c);
+                        puretriples.remove(c);
                     }
                 }
 
@@ -418,22 +450,18 @@ public class FindOneFactorClusters {
         }
 
         // Strict grow phase.
-        if (algType == AlgType.strict) {
+        if (false) {
             Set<Integer> t = new HashSet<Integer>();
             int count = 0;
-            int total = threeClusters.size();
+            int total = puretriples.size();
 
             do {
-                if (!threeClusters.iterator().hasNext()) {
+                if (!puretriples.iterator().hasNext()) {
                     break;
                 }
 
-                Set<Integer> cluster = threeClusters.iterator().next();
+                Set<Integer> cluster = puretriples.iterator().next();
                 Set<Integer> _cluster = new HashSet<Integer>(cluster);
-
-                if (extraShuffle) {
-                    Collections.shuffle(_variables);
-                }
 
                 VARIABLES:
                 for (int o : _variables) {
@@ -441,19 +469,23 @@ public class FindOneFactorClusters {
 
                     List<Integer> _cluster2 = new ArrayList<Integer>(_cluster);
 
-                    ChoiceGenerator gen = new ChoiceGenerator(_cluster2.size(), 2);
+                    ChoiceGenerator gen = new ChoiceGenerator(_cluster2.size(), 4);
                     int[] choice;
 
                     while ((choice = gen.next()) != null) {
                         int n1 = _cluster2.get(choice[0]);
                         int n2 = _cluster2.get(choice[1]);
+                        int n3 = _cluster2.get(choice[2]);
+                        int n4 = _cluster2.get(choice[3]);
 
                         t.clear();
                         t.add(n1);
                         t.add(n2);
+                        t.add(n3);
+                        t.add(n4);
                         t.add(o);
 
-                        if (!threeClusters.contains(t)) {
+                        if (!puretriples.contains(t)) {
                             continue VARIABLES;
                         }
 
@@ -478,14 +510,79 @@ public class FindOneFactorClusters {
                     t.add(n2);
                     t.add(n3);
 
-                    threeClusters.remove(t);
+                    puretriples.remove(t);
                 }
 
                 if (verbose) {
                     System.out.println("Grown " + (++count) + " of " + total + ": " + _cluster);
                 }
                 grown.add(_cluster);
-            } while (!threeClusters.isEmpty());
+            } while (!puretriples.isEmpty());
+        }
+
+        if (false) {
+            System.out.println("# pure triples = " + puretriples.size());
+
+            List<Set<Integer>> clusters = new LinkedList<Set<Integer>>(puretriples);
+            Set<Integer> t = new HashSet<Integer>();
+
+            I:
+            for (int i = 0; i < clusters.size(); i++) {
+                System.out.println("I = " + i);
+
+//                // remove "i" clusters that intersect with previous clusters.
+//                for (int k = 0; k < i - 1; k++) {
+//                    Set<Integer> ck = clusters.get(k);
+//                    Set<Integer> ci = clusters.get(i);
+//
+//                    if (ck == null) continue;
+//                    if (ci == null) continue;
+//
+//                    Set<Integer> cm = new HashSet<Integer>(ck);
+//                    cm.retainAll(ci);
+//
+//                    if (!cm.isEmpty()) {
+//                        clusters.remove(i);
+//                        i--;
+//                        continue I;
+//                    }
+//                }
+
+                J:
+                for (int j = i + 1; j < clusters.size(); j++) {
+                    Set<Integer> ci = clusters.get(i);
+                    Set<Integer> cj = clusters.get(j);
+
+                    if (ci == null) continue;
+                    if (cj == null) continue;
+
+                    Set<Integer> ck = new HashSet<Integer>(ci);
+                    ck.addAll(cj);
+
+                    List<Integer> cm = new ArrayList<Integer>(ck);
+
+                    ChoiceGenerator gen = new ChoiceGenerator(cm.size(), 3);
+                    int[] choice;
+
+                    while ((choice = gen.next()) != null) {
+                        t.clear();
+                        t.add(cm.get(choice[0]));
+                        t.add(cm.get(choice[1]));
+                        t.add(cm.get(choice[2]));
+
+                        if (!puretriples.contains(t)) {
+                            continue J;
+                        }
+                    }
+
+                    clusters.set(i, ck);
+                    clusters.remove(j);
+                    j--;
+                    System.out.println("Removing " + ci + ", " + cj + ", adding " + ck);
+                }
+            }
+
+            grown = new HashSet<Set<Integer>>(clusters);
         }
 
         // Optimized pick phase.
@@ -503,510 +600,113 @@ public class FindOneFactorClusters {
 
         List<Set<Integer>> list = new ArrayList<Set<Integer>>(grown);
 
-//        final Map<Set<Integer>, Double> pValues = new HashMap<Set<Integer>, Double>();
-//
-//        for (Set<Integer> o : grown) {
-//            pValues.put(o, getP(new ArrayList<Integer>(o)));
-//        }
-
         Collections.sort(list, new Comparator<Set<Integer>>() {
             @Override
             public int compare(Set<Integer> o1, Set<Integer> o2) {
-//                if (o1.size() == o2.size()) {
-//                    double chisq1 = pValues.get(o1);
-//                    double chisq2 = pValues.get(o2);
-//                    return Double.compare(chisq2, chisq1);
-//                }
-
                 return o2.size() - o1.size();
             }
         });
 
-//        for (Set<Integer> o : list) {
-//            if (pValues.get(o) < alpha) continue;
-//            System.out.println(variablesForIndices(new ArrayList<Integer>(o)) + "  p = " + pValues.get(o));
-//        }
+//        final Map<Set<Integer>, Double> significances = new HashMap<Set<Integer>, Double>();
+
+//        Collections.sort(list, new Comparator<Set<Integer>>() {
+//            @Override
+//            public int compare(Set<Integer> cluster1, Set<Integer> cluster2) {
+////                Double sum1 = significances.get(cluster1);
+////                if (sum1 == null) {
+////                    double sig = significance(new ArrayList<Integer>(cluster1));
+////                    significances.put(cluster1, sig);
+////                    sum1 = sig;
+////                }
+////                Double sum2 = significances.get(cluster2);
+////                if (sum2 == null) {
+////                    double sig = significance(new ArrayList<Integer>(cluster2));
+////                    significances.put(cluster2, sig);
+////                    sum2 = sig;
+////                }
+//
+//                double avg1 = avgSumLnP(new ArrayList<Integer>(cluster1));
+//                double avg2 = avgSumLnP(new ArrayList<Integer>(cluster2));
+//
+//                return Double.compare(avg2, avg1);
+//            }
+//        });
 
         Set<Integer> all = new HashSet<Integer>();
 
         CLUSTER:
         for (Set<Integer> cluster : list) {
-//            if (pValues.get(cluster) < alpha) continue;
-
             for (Integer i : cluster) {
                 if (all.contains(i)) continue CLUSTER;
             }
 
             out.add(cluster);
-
-//            if (getPMulticluster(out) < alpha) {
-//                out.remove(cluster);
-//                continue;
-//            }
-
             all.addAll(cluster);
+        }
+
+        if (significanceCalculated) {
+            for (Set<Integer> _out : out) {
+                try {
+                    double p = significance(new ArrayList<Integer>(_out));
+                    log("OUT: " + variablesForIndices(new ArrayList<Integer>(_out)) + " p = " + p, true);
+                } catch (Exception e) {
+                    log("OUT: " + variablesForIndices(new ArrayList<Integer>(_out)) + " p = EXCEPTION", true);
+                }
+            }
+        } else {
+            for (Set<Integer> _out : out) {
+                log("OUT: " + variablesForIndices(new ArrayList<Integer>(_out)), true);
+            }
         }
 
         return out;
     }
 
-    private double getP(List<Integer> o) {
-        if (o.size() == 3) return 0;
+    Map<Set<Integer>, Double> avgSumLnPs = new HashMap<Set<Integer>, Double>();
 
-        double max = Double.NEGATIVE_INFINITY;
-
-        for (int i = 0; i < 1; i++) {
-            double c = getP(new ArrayList<Integer>(o), 3);
-            if (c > max) max = c;
-        }
-
-        return max;
-    }
-
-    private double getP(List<Integer> cluster, int numRestarts) {
-        if (true) {
-            Node latent = new GraphNode("L");
-            latent.setNodeType(NodeType.LATENT);
-            Graph g = new EdgeListGraph();
-            g.addNode(latent);
-            List<Node> measures = variablesForIndices(cluster);
-            for (Node node : measures) {
-                g.addNode(node);
-                g.addDirectedEdge(latent, node);
-            }
-            SemPm pm = new SemPm(g);
-
-//            pm.fixOneLoadingPerLatent();
-
-            SemOptimizer semOptimizer = new SemOptimizerPowell();
-            semOptimizer.setNumRestarts(numRestarts);
-
-            SemEstimator est = new SemEstimator(cov, pm, semOptimizer);
-            est.setScoreType(SemIm.ScoreType.Fgls);
-            est.estimate();
-            return est.getEstimatedSem().getPValue();
-        } else {
-            double max = Double.NEGATIVE_INFINITY;
-
-            for (int i = 0; i < numRestarts; i++) {
-                Mimbuild2 mimbuild = new Mimbuild2();
-
-                List<List<Node>> clusters1 = new ArrayList<List<Node>>();
-                clusters1.add(variablesForIndices(new ArrayList<Integer>(cluster)));
-
-                List<String> names = new ArrayList<String>();
-                names.add("L");
-
-                mimbuild.search(clusters1, names, cov);
-
-                double c = mimbuild.getpValue();
-                if (c > max) max = c;
-            }
-
-            return max;
-        }
-    }
-
-    private double getPMulticluster(Set<Set<Integer>> o) {
-        List<List<Integer>> list = new ArrayList<List<Integer>>();
-
-        for (Set<Integer> _o : o) {
-            list.add(new ArrayList<Integer>(_o));
-        }
-        return getPMulticluster(list, 3);
-    }
-
-    private double getPMulticluster(List<List<Integer>> clusters, int numRestarts) {
-        if (false) {
-            Graph g = new EdgeListGraph();
-            List<Node> latents = new ArrayList<Node>();
-            for (int i = 0; i < clusters.size(); i++) {
-                GraphNode latent = new GraphNode("L" + i);
-                latent.setNodeType(NodeType.LATENT);
-                latents.add(latent);
-                g.addNode(latent);
-
-                List<Node> cluster = variablesForIndices(clusters.get(i));
-
-                for (int j = 0; j < cluster.size(); j++) {
-                    g.addNode(cluster.get(j));
-                    g.addDirectedEdge(latent, cluster.get(j));
-                }
-            }
-            SemPm pm = new SemPm(g);
-
-//            pm.fixOneLoadingPerLatent();
-
-            SemOptimizer semOptimizer = new SemOptimizerPowell();
-            semOptimizer.setNumRestarts(numRestarts);
-
-            SemEstimator est = new SemEstimator(cov, pm, semOptimizer);
-            est.setScoreType(SemIm.ScoreType.Fgls);
-            est.estimate();
-            return est.getEstimatedSem().getPValue();
-        } else {
-            double max = Double.NEGATIVE_INFINITY;
-
-            for (int i = 0; i < numRestarts; i++) {
-                Mimbuild2 mimbuild = new Mimbuild2();
-
-                List<List<Node>> _clusters = new ArrayList<List<Node>>();
-
-                for (List<Integer> _cluster : clusters) {
-                    _clusters.add(variablesForIndices(_cluster));
-                }
-
-                List<String> names = new ArrayList<String>();
-
-                for (int j = 0; j < clusters.size(); j++) {
-                    names.add("L" + j);
-                }
-
-                mimbuild.search(_clusters, names, cov);
-
-                double c = mimbuild.getpValue();
-                if (c > max) max = c;
-            }
-
-            return max;
-        }
-    }
-
-    private void log(String s, boolean toLog) {
-        if (toLog) {
-            TetradLogger.getInstance().log("info", s);
-        }
-
-//        System.out.println(s);
-    }
-
-//    private Set<Set<Integer>> findThreeClusters() {
-////        Graph graph = new EdgeListGraph(variables);
-////        Fas fas = new Fas(graph, indTest);
-////        fas.setDepth(0);     // 1?
-////        Map<Node, Set<Node>> adjacencies = fas.searchMapOnly();
-//
+    // Finds clusters of size 4 or higher for the tetrad first algorithm.
+    private Set<List<Integer>> findPureClusters(List<Integer> _variables) {
+        Set<List<Integer>> clusters = new HashSet<List<Integer>>();
 //        List<Integer> allVariables = new ArrayList<Integer>();
 //        for (int i = 0; i < this.variables.size(); i++) allVariables.add(i);
-//
-//        if (allVariables.size() < 4) {
-//            return new HashSet<Set<Integer>>();
-//        }
-//
-//        ChoiceGenerator gen = new ChoiceGenerator(allVariables.size(), 3);
-//        int[] choice;
-//        Set<Set<Integer>> threeClusters = new HashSet<Set<Integer>>();
-//
-//        CHOICE:
-//        while ((choice = gen.next()) != null) {
-//            int n1 = allVariables.get(choice[0]);
-//            int n2 = allVariables.get(choice[1]);
-//            int n3 = allVariables.get(choice[2]);
-//
-//            Set<Integer> triple = triple(n1, n2, n3);
-//
-////            if (!clique(triple, adjacencies)) {
-////                continue;
-////            }
-//
-//            int rejected = 0;
-//
-//            for (int o : allVariables) {
-//                if (triple.contains(o)) {
-//                    continue;
-//                }
-//
-//                Set<Integer> cluster = quartet(o, n1, n2, n3);
-//
-//                if (!quartetVanishes(cluster)) {
-//                    rejected++;
-//                }
-//
-//                if (rejected > variables.size() * 0.1) {
-//                    continue CHOICE;
-//                }
-////
-////                if (rejected > 0) {
-////                    continue CHOICE;
-////                }
-//            }
-//
-//            System.out.println("++ " + triple);
-//
-//            threeClusters.add(triple);
-//        }
-//
-//        return threeClusters;
-//    }
-
-//    private Set<Set<Integer>> combineThreeClusters(Set<Set<Integer>> threeClusters) {
-//        Set<Set<Integer>> grown = new HashSet<Set<Integer>>();
-//        List<Integer> _variables = new ArrayList<Integer>();
-//        for (int i = 0; i < variables.size(); i++) _variables.add(i);
-//
-//        for (Set<Integer> cluster : threeClusters) {
-//            Set<Integer> _cluster = new HashSet<Integer>(cluster);
-//
-//            // Strict.
-//            for (int o : _variables) {
-//                _cluster.add(o);
-//
-//                List<Integer> _cluster2 = new ArrayList<Integer>(_cluster);
-//
-//                ChoiceGenerator gen = new ChoiceGenerator(_cluster2.size(), 3);
-//                int[] choice;
-//                boolean rejected = false;
-//
-//                while ((choice = gen.next()) != null) {
-//                    int n1 = _cluster2.get(choice[0]);
-//                    int n2 = _cluster2.get(choice[1]);
-//                    int n3 = _cluster2.get(choice[2]);
-//
-//                    Set<Integer> triple = triple(n1, n2, n3);
-//
-//                    if (!triple.contains(o)) continue;
-//
-//                    if (!threeClusters.contains(triple)) {
-//                        rejected = true;
-//                        break;
-//                    }
-//                }
-//
-//                if (rejected) {
-//                    _cluster.remove(o);
-//                }
-//            }
-//
-//            grown.add(_cluster);
-//        }
-//
-//        List<Set<Integer>> _grown = new ArrayList<Set<Integer>>(grown);
-//        final Map<Set<Integer>, Double> pValues = new HashMap<Set<Integer>, Double>();
-//
-//        if (!Double.isNaN(clusterMinP) || sortKey == SortKey.pValue) {
-//            for (Set<Integer> g : _grown) {
-//                double p = getClusterP2(variablesForIndices(g));
-//                pValues.put(g, p);
-//            }
-//        }
-//
-//        // Print the grown clusters.
-//        Collections.sort(_grown, new Comparator<Set<Integer>>() {
-//            public int compare(Set<Integer> o1, Set<Integer> o2) {
-//                if (sortKey == SortKey.pValue) {
-//                    Double p1 = pValues.get(o2);
-//                    Double p2 = pValues.get(o1);
-//                    return Double.compare(Double.isNaN(p1) ? -1 : p1, Double.isNaN(p2) ? -1 : p2);
-//                } else if (sortKey == SortKey.size) {
-//                    return o2.size() - o1.size();
-//                } else {
-//                    throw new IllegalStateException();
-//                }
-//            }
-//        });
-//
-////        System.out.println("Grown");
-////        for (Set<Integer> l : _grown) {
-////            List<Node> nodes = variablesForIndices(l);
-////
-////            for (int i = 0; i < nodes.size(); i++) {
-////                System.out.print(nodes.get(i));
-////
-////                if (i < nodes.size() - 1) {
-////                    System.out.print(" ");
-////                }
-////            }
-////
-////            if (sortKey == SortKey.size) {
-////                System.out.println();
-////            }
-////            if (sortKey == SortKey.pValue) {
-////                System.out.println("\t" + pValues.get(l));
-////            }
-////        }
-//
-//        Set<Set<Integer>> out = new HashSet<Set<Integer>>();
-//
-//        while (!_grown.isEmpty()) {
-//            Set<Integer> maxCluster = _grown.remove(0);
-//            if (!Double.isNaN(clusterMinP) && maxCluster.size() == 3) {
-//                _grown.remove(maxCluster);
-//                continue;
-//            }
-//            if (!Double.isNaN(clusterMinP) && pValues.get(maxCluster) < clusterMinP) {
-//                grown.remove(maxCluster);
-//                continue;
-//            }
-//            out.add(maxCluster);
-//
-//            // Remove from grown any cluster that intersects it.
-//            for (Set<Integer> _cluster : new HashSet<Set<Integer>>(_grown)) {
-//                Set<Integer> cluster2 = new HashSet<Integer>(_cluster);
-//                cluster2.retainAll(maxCluster);
-//
-//                if (!cluster2.isEmpty()) {
-//                    _grown.remove(_cluster);
-//                }
-//            }
-//        }
-//
-////        NumberFormat nf = new DecimalFormat("0.0000");
-////
-////        // Print the output clusters.
-////        System.out.println("Output clusters:");
-////
-////        for (Set<Integer> l : out) {
-////            List<Node> nodes = variablesForIndices(l);
-////
-////            for (int i = 0; i < nodes.size(); i++) {
-////                System.out.print(nodes.get(i));
-////
-////                if (i < nodes.size() - 1) {
-////                    System.out.print(" ");
-////                }
-////            }
-////
-////            if (sortKey == SortKey.size) {
-////                System.out.println();
-////            }
-////            else if (sortKey == SortKey.pValue) {
-////                System.out.println("\t" + nf.format(pValues.get(l)));
-////            }
-////        }
-//
-//        return out;
-//    }
-
-    // Quartets first , then triples.
-    private Set<Set<Integer>> estimateClustersSAG() {
-        Map<Node, Set<Node>> adjacencies;
-
-        if (depth == -2) {
-            adjacencies = new HashMap<Node, Set<Node>>();
-
-            for (Node node : variables) {
-                HashSet<Node> _nodes = new HashSet<Node>(variables);
-                _nodes.remove(node);
-                adjacencies.put(node, _nodes);
-            }
-        } else {
-            System.out.println("Running PC adjacency search...");
-            Graph graph = new EdgeListGraph(variables);
-            Fas fas = new Fas(graph, indTest);
-            fas.setDepth(depth);     // 1?
-            adjacencies = fas.searchMapOnly();
-            System.out.println("...done.");
-        }
-
-        List<Integer> _variables = new ArrayList<Integer>();
-        for (int i = 0; i < variables.size(); i++) _variables.add(i);
-
-        Set<Set<Integer>> pureClusters = findPureClusters(_variables, adjacencies);
-        for (Set<Integer> cluster : pureClusters) _variables.removeAll(cluster);
-        Set<Set<Integer>> mixedClusters = findMixedClusters(_variables, unionPure(pureClusters), adjacencies);
-        Set<Set<Integer>> allClusters = new HashSet<Set<Integer>>(pureClusters);
-        allClusters.addAll(mixedClusters);
-        return allClusters;
-
-    }
-
-    // Finds clusters of size 4 or higher.
-    private Set<Set<Integer>> findPureClusters(List<Integer> _variables, Map<Node, Set<Node>> adjacencies) {
-//        System.out.println("Original variables = " + variables);
-
-        Set<Set<Integer>> clusters = new HashSet<Set<Integer>>();
-        List<Integer> allVariables = new ArrayList<Integer>();
-        for (int i = 0; i < this.variables.size(); i++) allVariables.add(i);
+        List<Integer> allVariables = allVariables();
 
         VARIABLES:
         while (!_variables.isEmpty()) {
+            if (verbose) {
+                System.out.println(_variables);
+            }
             if (_variables.size() < 4) break;
 
-            for (int x : _variables) {
-                Node nodeX = variables.get(x);
-                List<Node> adjX = new ArrayList<Node>(adjacencies.get(nodeX));
-                adjX.retainAll(variablesForIndices(new ArrayList<Integer>(_variables)));
+            ChoiceGenerator gen = new ChoiceGenerator(_variables.size(), 4);
+            int[] choice;
 
-                for (Node node : new ArrayList<Node>(adjX)) {
-                    if (adjacencies.get(node).size() < 3) {
-                        adjX.remove(node);
-                    }
-                }
+            while ((choice = gen.next()) != null) {
+                int n1 = _variables.get(choice[0]);
+                int n2 = _variables.get(choice[1]);
+                int n3 = _variables.get(choice[2]);
+                int n4 = _variables.get(choice[3]);
 
-                if (adjX.size() < 3) {
-                    continue;
-                }
+                List<Integer> cluster = quartet(n1, n2, n3, n4);
 
-                ChoiceGenerator gen = new ChoiceGenerator(adjX.size(), 3);
-                int[] choice;
-
-                while ((choice = gen.next()) != null) {
-                    Node nodeY = adjX.get(choice[0]);
-                    Node nodeZ = adjX.get(choice[1]);
-                    Node nodeW = adjX.get(choice[2]);
-
-                    int y = variables.indexOf(nodeY);
-                    int w = variables.indexOf(nodeW);
-                    int z = variables.indexOf(nodeZ);
-
-                    Set<Integer> cluster = quartet(x, y, z, w);
-
-                    if (!clique(cluster, adjacencies)) {
-                        continue;
+                // Note that purity needs to be assessed with respect to all of the variables in order to
+                // remove all latent-measure impurities between pairs of latents.
+                if (pure(cluster, allVariables, alpha)) {
+                    if (verbose) {
+                        log("Found a pure: " + variablesForIndices(cluster), false);
                     }
 
-                    // Note that purity needs to be assessed with respect to all of the variables in order to
-                    // remove all latent-measure impurities between pairs of latents.
-                    if (pure(cluster, allVariables)) {
+//                    if (modelInsignificantWithNewCluster(clusters, cluster)) continue;
 
-//                        Collections.shuffle(_variables);
+                    addOtherVariables(_variables, allVariables, cluster);
 
-                        O:
-                        for (int o : _variables) {
-                            if (cluster.contains(o)) continue;
-                            cluster.add(o);
-                            List<Integer> _cluster = new ArrayList<Integer>(cluster);
-
-                            if (!clique(cluster, adjacencies)) {
-                                cluster.remove(o);
-                                continue O;
-                            }
-
-//                            if (!allVariablesDependent(cluster)) {
-//                                cluster.remove(o);
-//                                continue O;
-//                            }
-
-                            ChoiceGenerator gen2 = new ChoiceGenerator(_cluster.size(), 4);
-                            int[] choice2;
-                            int count = 0;
-
-                            while ((choice2 = gen2.next()) != null) {
-                                int x2 = _cluster.get(choice2[0]);
-                                int y2 = _cluster.get(choice2[1]);
-                                int z2 = _cluster.get(choice2[2]);
-                                int w2 = _cluster.get(choice2[3]);
-
-                                Set<Integer> quartet = quartet(x2, y2, z2, w2);
-
-                                // Optimizes for large clusters.
-                                if (quartet.contains(o)) {
-                                    if (++count > 50) continue O;
-                                }
-
-                                if (quartet.contains(o) && !pure(quartet, allVariables)) {
-                                    cluster.remove(o);
-                                    continue O;
-                                }
-                            }
-                        }
-
-                        System.out.println("Cluster found: " + variablesForIndices(new ArrayList<Integer>(cluster)));
-                        clusters.add(cluster);
-                        _variables.removeAll(cluster);
-
-                        continue VARIABLES;
+                    if (verbose) {
+                        log("Cluster found: " + variablesForIndices(cluster), true);
                     }
+                    clusters.add(cluster);
+                    _variables.removeAll(cluster);
+
+                    continue VARIABLES;
                 }
             }
 
@@ -1016,160 +716,122 @@ public class FindOneFactorClusters {
         return clusters;
     }
 
-    // Trying to optimize the search for 4-cliques a bit.
-    private Set<Set<Integer>> findPureClusters2(List<Integer> _variables, Map<Node, Set<Node>> adjacencies) {
-        System.out.println("Original variables = " + variables);
+    private void addOtherVariables(List<Integer> _variables, List<Integer> allVariables, List<Integer> cluster) {
+        O:
+        for (int o : _variables) {
+            if (cluster.contains(o)) continue;
+            List<Integer> _cluster = new ArrayList<Integer>(cluster);
 
-        Set<Set<Integer>> clusters = new HashSet<Set<Integer>>();
-        List<Integer> allVariables = new ArrayList<Integer>();
-        Set<Node> foundVariables = new HashSet<Node>();
-        for (int i = 0; i < this.variables.size(); i++) allVariables.add(i);
+            ChoiceGenerator gen2 = new ChoiceGenerator(_cluster.size(), 3);
+            int[] choice2;
+//            boolean found = false;
 
-        for (int x : _variables) {
-            Node nodeX = variables.get(x);
-            if (foundVariables.contains(nodeX)) continue;
+            while ((choice2 = gen2.next()) != null) {
+                int t1 = _cluster.get(choice2[0]);
+                int t2 = _cluster.get(choice2[1]);
+                int t3 = _cluster.get(choice2[2]);
 
-            List<Node> adjX = new ArrayList<Node>(adjacencies.get(nodeX));
-            adjX.removeAll(foundVariables);
+                List<Integer> quartet = triple(t1, t2, t3);
 
-            if (adjX.size() < 3) continue;
 
-            for (Node nodeY : adjX) {
-                if (foundVariables.contains(nodeY)) continue;
+                quartet.add(o);
 
-                List<Node> commonXY = new ArrayList<Node>(adjacencies.get(nodeY));
-                commonXY.retainAll(adjX);
-                commonXY.removeAll(foundVariables);
+//                if (pure(quartet, allVariables, alpha)) {
+//                    found = true;
+//                    break;
+//                }
 
-                for (Node nodeZ : commonXY) {
-                    if (foundVariables.contains(nodeZ)) continue;
-
-                    List<Node> commonXZ = new ArrayList<Node>(commonXY);
-                    commonXZ.retainAll(adjacencies.get(nodeZ));
-                    commonXZ.removeAll(foundVariables);
-
-                    for (Node nodeW : commonXZ) {
-                        if (foundVariables.contains(nodeW)) continue;
-
-                        if (!adjacencies.get(nodeY).contains(nodeW)) {
-                            continue;
-                        }
-
-                        int y = variables.indexOf(nodeY);
-                        int w = variables.indexOf(nodeW);
-                        int z = variables.indexOf(nodeZ);
-
-                        Set<Integer> cluster = quartet(x, y, z, w);
-
-                        // Note that purity needs to be assessed with respect to all of the variables in order to
-                        // remove all latent-measure impurities between pairs of latents.
-                        if (pure(cluster, allVariables)) {
-
-                            O:
-                            for (int o : _variables) {
-                                if (cluster.contains(o)) continue;
-                                cluster.add(o);
-
-                                if (!clique(cluster, adjacencies)) {
-                                    cluster.remove(o);
-                                    continue O;
-                                }
-
-//                                if (!allVariablesDependent(cluster)) {
-//                                    cluster.remove(o);
-//                                    continue O;
-//                                }
-
-                                List<Integer> _cluster = new ArrayList<Integer>(cluster);
-
-                                ChoiceGenerator gen2 = new ChoiceGenerator(_cluster.size(), 4);
-                                int[] choice2;
-                                int count = 0;
-
-                                while ((choice2 = gen2.next()) != null) {
-                                    int x2 = _cluster.get(choice2[0]);
-                                    int y2 = _cluster.get(choice2[1]);
-                                    int z2 = _cluster.get(choice2[2]);
-                                    int w2 = _cluster.get(choice2[3]);
-
-                                    Set<Integer> quartet = quartet(x2, y2, z2, w2);
-
-                                    // Optimizes for large clusters.
-                                    if (quartet.contains(o)) {
-                                        if (++count > 2) continue O;
-                                    }
-
-                                    if (quartet.contains(o) && !pure(quartet, allVariables)) {
-                                        cluster.remove(o);
-                                        continue O;
-                                    }
-                                }
-                            }
-
-                            System.out.println("Cluster found: " + variablesForIndices(new ArrayList<Integer>(cluster)));
-                            clusters.add(cluster);
-                            foundVariables.addAll(variablesForIndices(new ArrayList<Integer>(cluster)));
-                        }
-                    }
+                if (!pure(quartet, allVariables, alpha)) {
+                    continue O;
                 }
             }
-        }
 
-        return clusters;
+//            if (found) {
+                log("Extending by " + variables.get(o), false);
+                cluster.add(o);
+//            }
+        }
     }
 
-    //  Finds clusters of size 3.
-    private Set<Set<Integer>> findMixedClusters(List<Integer> remaining, Set<Integer> unionPure, Map<Node, Set<Node>> adjacencies) {
-        Set<Set<Integer>> threeClusters = new HashSet<Set<Integer>>();
+    private boolean modelInsignificantWithNewCluster(Set<List<Integer>> clusters, List<Integer> cluster) {
+//        if (true) return false;
+
+        List<List<Integer>> __clusters = new ArrayList<List<Integer>>(clusters);
+        __clusters.add(cluster);
+        double significance3 = getModelPValue(__clusters);
+        if (verbose) {
+            log("Significance * " + __clusters + " = " + significance3, false);
+        }
+
+        return significance3 < alpha;
+    }
+
+    //  Finds clusters of size 3 3or the quartet first algorithm.
+    private Set<List<Integer>> findMixedClusters(Set<List<Integer>> clusters, List<Integer> remaining, Set<Integer> unionPure) {
+        Set<List<Integer>> triples = new HashSet<List<Integer>>();
+        Set<List<Integer>> _clusters = new HashSet<List<Integer>>(clusters);
 
         if (unionPure.isEmpty()) {
-            return new HashSet<Set<Integer>>();
+            return new HashSet<List<Integer>>();
         }
 
         REMAINING:
         while (true) {
             if (remaining.size() < 3) break;
 
+            if (verbose) {
+                log("UnionPure = " + variablesForIndices(new ArrayList<Integer>(unionPure)), false);
+            }
+
             ChoiceGenerator gen = new ChoiceGenerator(remaining.size(), 3);
             int[] choice;
 
             while ((choice = gen.next()) != null) {
-                int y = remaining.get(choice[0]);
-                int z = remaining.get(choice[1]);
-                int w = remaining.get(choice[2]);
+                int t2 = remaining.get(choice[0]);
+                int t3 = remaining.get(choice[1]);
+                int t4 = remaining.get(choice[2]);
 
-                Set<Integer> cluster = new HashSet<Integer>();
-                cluster.add(y);
-                cluster.add(z);
-                cluster.add(w);
+                List<Integer> cluster = new ArrayList<Integer>();
+                cluster.add(t2);
+                cluster.add(t3);
+                cluster.add(t4);
 
-//                if (!allVariablesDependent(cluster)) {
-//                    continue;
-//                }
-
-                if (!clique(cluster, adjacencies)) {
+                if (zeroCorr(cluster)) {
                     continue;
                 }
 
                 // Check all x as a cross check; really only one should be necessary.
-                boolean allX = true;
+                boolean allvanish = true;
+                boolean someVanish = false;
 
-                for (int x : unionPure) {
-                    Set<Integer> _cluster = new HashSet<Integer>(cluster);
-                    _cluster.add(x);
+                for (int t1 : allVariables()) {
+                    if (cluster.contains(t1)) continue;
 
-                    if (!quartetVanishes(_cluster) || !significant(new ArrayList<Integer>(_cluster))) {
-                        allX = false;
+                    List<Integer> _cluster = new ArrayList<Integer>(cluster);
+                    _cluster.add(t1);
+
+
+                    if (vanishes(_cluster)) {
+//                        System.out.println("Vanishes: " + variablesForIndices(_cluster));
+                        someVanish = true;
+                    } else {
+//                        System.out.println("Doesn't vanish: " + variablesForIndices(_cluster));
+                        allvanish = false;
                         break;
                     }
                 }
 
-                if (allX) {
-                    threeClusters.add(cluster);
+                if (someVanish && allvanish) {
+//                    if (modelInsignificantWithNewCluster(_clusters, cluster)) continue;
+
+                    triples.add(cluster);
+                    _clusters.add(cluster);
                     unionPure.addAll(cluster);
                     remaining.removeAll(cluster);
 
-                    System.out.println("3-cluster found: " + variablesForIndices(new ArrayList<Integer>(cluster)));
+                    if (verbose) {
+                        log("3-cluster found: " + variablesForIndices(cluster), false);
+                    }
 
                     continue REMAINING;
                 }
@@ -1178,24 +840,33 @@ public class FindOneFactorClusters {
             break;
         }
 
-        return threeClusters;
+        return triples;
     }
 
-    private boolean clique(Set<Integer> cluster, Map<Node, Set<Node>> adjacencies) {
-        List<Integer> _cluster = new ArrayList<Integer>(cluster);
+    private double significance(List<Integer> cluster) {
+        double chisq = getClusterChiSquare(cluster);
 
-        for (int i = 0; i < cluster.size(); i++) {
-            for (int j = i + 1; j < cluster.size(); j++) {
-                Node nodei = variables.get(_cluster.get(i));
-                Node nodej = variables.get(_cluster.get(j));
+        // From "Algebraic factor analysis: tetrads, triples and beyond" Drton et al.
+        int n = cluster.size();
+        int dof = dofHarman(n);
+        double q = ProbUtils.chisqCdf(chisq, dof);
+        return 1.0 - q;
+    }
 
-                if (!adjacencies.get(nodei).contains(nodej)) {
-                    return false;
-                }
-            }
-        }
+    private double modelSignificance(List<List<Integer>> clusters) {
+        return getModelPValue(clusters);
+    }
 
-        return true;
+    private int dofDrton(int n) {
+        int dof = ((n - 2) * (n - 3)) / 2 - 2;
+        if (dof < 0) dof = 0;
+        return dof;
+    }
+
+    private int dofHarman(int n) {
+        int dof = n * (n - 5) / 2 + 1;
+        if (dof < 0) dof = 0;
+        return dof;
     }
 
     private List<Node> variablesForIndices(List<Integer> cluster) {
@@ -1205,68 +876,229 @@ public class FindOneFactorClusters {
             _cluster.add(variables.get(c));
         }
 
-        Collections.sort(_cluster);
+//        Collections.sort(_cluster);
 
         return _cluster;
     }
 
+    private List<List<Node>> variablesForIndices2(Set<List<Integer>> clusters) {
+        List<List<Node>> variables = new ArrayList<List<Node>>();
 
-    private boolean pure(Set<Integer> quartet, List<Integer> variables) {
-        if (quartetVanishes(quartet)) {
-            for (int o : variables) {
+        for (List<Integer> cluster : clusters) {
+            variables.add(variablesForIndices(cluster));
+        }
+
+        return variables;
+    }
+
+    private boolean pure(List<Integer> quartet, List<Integer> variables, double alpha) {
+        if (zeroCorr(quartet)) {
+            return false;
+        }
+
+        if (vanishes(quartet)) {
+            for (int o : allVariables()) {
                 if (quartet.contains(o)) continue;
 
-                for (int p : quartet) {
-                    Set<Integer> _quartet = new HashSet<Integer>(quartet);
-                    _quartet.remove(p);
+                for (int i = 0; i < quartet.size(); i++) {
+                    List<Integer> _quartet = new ArrayList<Integer>(quartet);
+                    _quartet.remove(quartet.get(i));
                     _quartet.add(o);
 
-                    if (!quartetVanishes(_quartet)) {
+//                    if (zeroCorr(_quartet)) {
+//                        continue;
+//                    }
+
+                    if (!(vanishes(_quartet))) {
                         return false;
                     }
                 }
             }
 
-            return significant(new ArrayList<Integer>(quartet));
+            return true;
         }
 
         return false;
     }
 
-    private Set<Integer> quartet(int x, int y, int z, int w) {
-        Set<Integer> set = new HashSet<Integer>();
-        set.add(x);
-        set.add(y);
-        set.add(z);
-        set.add(w);
-
-        if (set.size() < 4)
-            throw new IllegalArgumentException("Quartet elements must be unique: <" + x + ", " + y + ", " + z + ", " + w + ">");
-
-        return set;
+    private double getClusterChiSquare(List<Integer> cluster) {
+        SemIm im = estimateClusterModel(cluster);
+        return im.getChiSquare();
     }
 
-    private boolean quartetVanishes(Set<Integer> quartet) {
-        if (quartet.size() != 4) throw new IllegalArgumentException("Expecting a quartet, size = " + quartet.size());
+    private SemIm estimateClusterModel(List<Integer> quartet) {
+        Graph g = new EdgeListGraph();
+        Node l1 = new GraphNode("L1");
+        l1.setNodeType(NodeType.LATENT);
+        Node l2 = new GraphNode("L2");
+        l2.setNodeType(NodeType.LATENT);
+        g.addNode(l1);
+        g.addNode(l2);
 
-        Iterator<Integer> iter = quartet.iterator();
-        int x = iter.next();
-        int y = iter.next();
-        int z = iter.next();
-        int w = iter.next();
+        for (int i = 0; i < quartet.size(); i++) {
+            Node n = this.variables.get(quartet.get(i));
+            g.addNode(n);
+            g.addDirectedEdge(l1, n);
+            g.addDirectedEdge(l2, n);
+        }
 
-        return testVanishing(x, y, z, w);
+        SemPm pm = new SemPm(g);
+
+        SemEstimator est;
+
+        if (dataModel instanceof DataSet) {
+            est = new SemEstimator((DataSet) dataModel, pm, new SemOptimizerEm());
+        } else {
+            est = new SemEstimator((CovarianceMatrix) dataModel, pm, new SemOptimizerEm());
+        }
+
+        return est.estimate();
     }
 
-    private boolean testVanishing(int x, int y, int z, int w) {
+    private double getModelPValue(List<List<Integer>> clusters) {
+        SemIm im = estimateModel(clusters);
+        return im.getPValue();
+    }
+
+    private SemIm estimateModel(List<List<Integer>> clusters) {
+        Graph g = new EdgeListGraph();
+
+        List<Node> upperLatents = new ArrayList<Node>();
+        List<Node> lowerLatents = new ArrayList<Node>();
+
+        for (int i = 0; i < clusters.size(); i++) {
+            List<Integer> cluster = clusters.get(i);
+            Node l1 = new GraphNode("L1." + (i + 1));
+            l1.setNodeType(NodeType.LATENT);
+
+            Node l2 = new GraphNode("L2." + (i + 1));
+            l2.setNodeType(NodeType.LATENT);
+
+            upperLatents.add(l1);
+            lowerLatents.add(l2);
+
+            g.addNode(l1);
+            g.addNode(l2);
+
+            for (int k = 0; k < cluster.size(); k++) {
+                Node n = this.variables.get(cluster.get(k));
+                g.addNode(n);
+                g.addDirectedEdge(l1, n);
+                g.addDirectedEdge(l2, n);
+            }
+        }
+
+        for (int i = 0; i < upperLatents.size(); i++) {
+            for (int j = i + 1; j < upperLatents.size(); j++) {
+                g.addDirectedEdge(upperLatents.get(i), upperLatents.get(j));
+                g.addDirectedEdge(lowerLatents.get(i), lowerLatents.get(j));
+            }
+        }
+
+        for (int i = 0; i < upperLatents.size(); i++) {
+            for (int j = 0; j < lowerLatents.size(); j++) {
+                if (i == j) continue;
+                g.addDirectedEdge(upperLatents.get(i), lowerLatents.get(j));
+            }
+        }
+
+        SemPm pm = new SemPm(g);
+
+        for (Node node : upperLatents) {
+            Parameter p = pm.getParameter(node, node);
+            p.setFixed(true);
+            p.setStartingValue(1.0);
+        }
+
+        for (Node node : lowerLatents) {
+            Parameter p = pm.getParameter(node, node);
+            p.setFixed(true);
+            p.setStartingValue(1.0);
+        }
+
+        SemEstimator est;
+
+        if (dataModel instanceof DataSet) {
+            est = new SemEstimator((DataSet) dataModel, pm, new SemOptimizerEm());
+        } else {
+            est = new SemEstimator((CovarianceMatrix) dataModel, pm, new SemOptimizerEm());
+        }
+
+        return est.estimate();
+    }
+
+    private List<Integer> quartet(int n1, int n2, int n3, int n4) {
+        List<Integer> quartet = new ArrayList<Integer>();
+        quartet.add(n1);
+        quartet.add(n2);
+        quartet.add(n3);
+        quartet.add(n4);
+
+        if (new HashSet<Integer>(quartet).size() < 4)
+            throw new IllegalArgumentException("quartet elements must be unique: <" + n1 + ", " + n2 + ", " + n3 + ", " + n4 + ">");
+
+        return quartet;
+    }
+
+    private List<Integer> triple(int n1, int n2, int n3) {
+        List<Integer> triple = new ArrayList<Integer>();
+        triple.add(n1);
+        triple.add(n2);
+        triple.add(n3);
+
+        if (new HashSet<Integer>(triple).size() < 3)
+            throw new IllegalArgumentException("triple elements must be unique: <" + n1 + ", " + n2 + ", " + n3 + ">");
+
+        return triple;
+    }
+
+    private boolean vanishes(List<Integer> quartet) {
+        int n1 = quartet.get(0);
+        int n2 = quartet.get(1);
+        int n3 = quartet.get(2);
+        int n4 = quartet.get(3);
+
+        return vanishes(n1, n2, n3, n4);
+    }
+
+    private boolean zeroCorr(List<Integer> cluster) {
+        int count = 0;
+
+        for (int i = 0; i < cluster.size(); i++) {
+            for (int j = i + 1; j < cluster.size(); j++) {
+                double r = this.corr.getValue(cluster.get(i), cluster.get(j));
+                int N = this.corr.getSampleSize();
+                double f = sqrt(N) * Math.log((1. + r) / (1. - r));
+                double p = 2.0 * (1.0 - RandomUtil.getInstance().normalCdf(0, 1, abs(f)));
+                if (p > alpha) count++;
+            }
+        }
+
+        return count >= 1;
+    }
+
+    /**
+     * The clusters output by the algorithm from the last call to search().
+     */
+    public List<List<Node>> getClusters() {
+        return clusters;
+    }
+
+    public void setVerbose(boolean verbose) {
+        this.verbose = verbose;
+    }
+
+    private boolean vanishes(int x, int y, int z, int w) {
         if (testType == TestType.TETRAD_DELTA) {
             Tetrad t1 = new Tetrad(variables.get(x), variables.get(y), variables.get(z), variables.get(w));
             Tetrad t2 = new Tetrad(variables.get(x), variables.get(y), variables.get(w), variables.get(z));
-            double p = deltaTest.getPValue(t1, t2);
-            return p > alpha;
-        } else {
-            return test.tetradHolds(x, y, z, w) && test.tetradHolds(x, y, w, z);
+
+            return test.getPValue(t1, t2) > alpha;
+        } else if (testType == TestType.TETRAD_WISHART) {
+            return test2.tetradPValue(x, y, z, w) > alpha && test2.tetradPValue(x, y, w, z) > alpha;
         }
+
+        throw new IllegalArgumentException("Only the delta and wishart tests are being used: " + testType);
     }
 
     private Graph convertSearchGraphNodes(Set<Set<Node>> clusters) {
@@ -1292,10 +1124,10 @@ public class FindOneFactorClusters {
         return graph;
     }
 
-    private Graph convertToGraph(Set<Set<Integer>> allClusters) {
+    private Graph convertToGraph(Set<List<Integer>> allClusters) {
         Set<Set<Node>> _clustering = new HashSet<Set<Node>>();
 
-        for (Set<Integer> cluster : allClusters) {
+        for (List<Integer> cluster : allClusters) {
             Set<Node> nodes = new HashSet<Node>();
 
             for (int i : cluster) {
@@ -1308,79 +1140,30 @@ public class FindOneFactorClusters {
         return convertSearchGraphNodes(_clustering);
     }
 
-    private Set<Integer> unionPure(Set<Set<Integer>> pureClusters) {
+    private Set<Integer> unionPure(Set<List<Integer>> pureClusters) {
         Set<Integer> unionPure = new HashSet<Integer>();
 
-        for (Set<Integer> cluster : pureClusters) {
+        for (List<Integer> cluster : pureClusters) {
             unionPure.addAll(cluster);
         }
 
         return unionPure;
     }
 
-    private boolean significant(List<Integer> cluster) {
-        double p = getClusterP2(variablesForIndices(new ArrayList<Integer>(cluster)));
-
-        return p > alpha;
-    }
-
-    private double getClusterP2(List<Node> c) {
-        Graph g = new EdgeListGraph(c);
-        Node l = new GraphNode("L");
-        l.setNodeType(NodeType.LATENT);
-        g.addNode(l);
-
-        for (Node n : c) {
-            g.addDirectedEdge(l, n);
+    private void log(String s, boolean toLog) {
+        if (toLog) {
+            TetradLogger.getInstance().log("info", s);
         }
 
-        SemPm pm = new SemPm(g);
-        SemEstimator est;
-        if (dataModel instanceof DataSet) {
-            est = new SemEstimator((DataSet) dataModel, pm, new SemOptimizerEm());
-        } else {
-            est = new SemEstimator((CovarianceMatrix) dataModel, pm, new SemOptimizerEm());
-        }
-        SemIm estIm = est.estimate();
-        double pValue = estIm.getPValue();
-        return pValue == 1 ? Double.NaN : pValue;
+        System.out.println(s);
     }
 
-
-    private Set<Integer> triple(int n1, int n2, int n3) {
-        Set<Integer> triple = new HashSet<Integer>();
-        triple.add(n1);
-        triple.add(n2);
-        triple.add(n3);
-
-        if (triple.size() < 3)
-            throw new IllegalArgumentException("Triple elements must be unique: <" + n1 + ", " + n2 + ", " + n3 + ">");
-
-        return triple;
+    public boolean isSignificanceCalculated() {
+        return significanceCalculated;
     }
 
-    public List<List<Node>> getClusters() {
-        return clusters;
-    }
-
-    public SortKey getSortKey() {
-        return sortKey;
-    }
-
-    public void setSortKey(SortKey sortKey) {
-        this.sortKey = sortKey;
-    }
-
-    /**
-     * Clusters with p value balow this will not be returned, saving you time.
-     */
-    public Double getClusterMinP() {
-        return clusterMinP;
-    }
-
-    public void setClusterMinP(Double clusterMinP) {
-        if (clusterMinP < 0 || clusterMinP > 1) throw new IllegalArgumentException();
-        this.clusterMinP = clusterMinP;
+    public void setSignificanceCalculated(boolean significanceCalculated) {
+        this.significanceCalculated = significanceCalculated;
     }
 }
